requirements.txt
coverage==7.4.0
django==4.2.9
pylint==2.7.0
django-extensions==3.2.3
djangorestframework==3.14.0
pylint==2.7.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-xdist==3.5.0


